Log number of parameters of a model
Background
We currently include the size information in the model name and use it to differentiate between model sizes.
However, I think it's too hard to tell what size this model is with a size name like '-base', '-large', '-b0' etc.

Even, nowadays, the number of parameters, such as 7b, is used as a suffix to express the size of the model.
This numbering tells us more information. How much GPU memory usage will be occupied when a model is loaded.
It also makes it easier to compare model sizes between similar types of models, such as trasnformers models.

What this PR does
calculate the number of parameters of a model and log using mlflow.log_param().